+++ web/config/default.settings.php
@@ -33,6 +33,9 @@ $app['csrf.secret'] = 'lkjihgfedcba';
 // Log path
 $app['log.path'] = __DIR__.'/../var/log/';
 
+// Logging level
+$app['log.level'] = 'INFO';
+
 // Base URL
 $app['caldav.baseurl'] = 'http://localhost:81/';