+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - Added new caldav.connect.timeout and caldav.response.timeout settings
 - Added a new preference to show a marker indicating the current time
+- Added a new `log.level` setting
 
 ## [2.0.0] - 2016-11-19